requirements.txt
discord.py>=2.3.2
openai>=1.3.0
APScheduler>=3.10.4
feedparser>=6.0.10
newspaper4k>=0.9.2
lxml[html_clean]>=4.9.0
beautifulsoup4>=4.12.2
python-dotenv>=1.0.0
requests>=2.31.0